requirements.txt
numpy
pandas
matplotlib
seaborn
scipy
statsmodels
wordcloud
pymysql
gradio
streamlit
langchain-community
scikit-learn
ydata-profiling
tensorflow
sqlalchemy
ipykernel
xgboost
joblib
lifelines
biopython
certifi